added shebang to rb

+++ Makefile
@@ -4,6 +4,7 @@ GO=$(EXECUTABLE).go
 JAVA=$(EXECUTABLE).java
 PY=$(EXECUTABLE).py
 RKT=$(EXECUTABLE).rkt
+RB=$(EXECUTABLE).rb
 
 help:
 	@echo 'make help:                this message'
@@ -12,10 +13,8 @@ help:
 	@echo 'make java:                compile to java class'
 	@echo 'make py:                  change python script permissions to executable'
 	@echo 'make rkt                  compile racket binary'
-	@echo 'make c install:           install c version'
-	@echo 'make go install:          install go version'
-	@echo 'make py install:          install python version'
-	@echo 'make rkt install:         install racket executable version'
+	@echo 'make rb:                  change ruby script permissions to executable'
+	@echo 'make install:             install to /usr/bin"
 
 c: $(C)
 	gcc $(C) -o $(EXECUTABLE)
@@ -33,10 +32,14 @@ py: $(PY)
 rkt: $(RKT)
 	raco exe $(RKT)
 
+rb: $(RB)
+	cp $(RB) $(EXECUTABLE)
+	chmod 744 $(EXECUTABLE)
+
 install: $EXECUTABLE)
 	mv $(EXECUTABLE) /usr/bin/
 
 clean: $(EXECUTABLE)
-	rm $(EXECUTABLE)
+	if [ -e $(EXECUTABLE) ]; then rm $(EXECUTABLE); fi
 	if [ -e "time2sex.class" ]; then rm time2sex.class; fi
